Reorganize & nest routes under App
